requirements.txt
# NDTA News Pipeline - Python Dependencies

# Core dependencies
python-dotenv==1.0.0
pyyaml==6.0.1

# Web scraping
feedparser==6.0.10
requests==2.31.0
beautifulsoup4==4.12.2

# AI/OpenAI
openai==0.28.1

# Social media APIs
tweepy==4.14.0

# Graphics generation
Pillow==10.1.0

# Utilities
python-dateutil==2.8.2

